+++ tests/unit/test_attribute_mapping_collection.py
@@ -39,6 +39,18 @@ class TestAttributeMappingCollection(TestCase):
         )
 
     @responses.activate
+    def test_delete(self):
+        general_url = (
+            "http://localhost:9100/api/versioned/v1/projects/4/attributeMappings"
+        )
+        responses.add(responses.GET, general_url, json=self.mappings_json)
+        deleteColl = AttributeMappingCollection(self.unify, general_url)
+        specific_url = general_url + "/19629-12"
+        responses.add(responses.DELETE, specific_url, status=204)
+        response = deleteColl.delete_by_resource_id("19629-12")
+        self.assertEqual(response.status_code, 204)
+
+    @responses.activate
     def test_create(self):
         def create_callback(request, snoop):
             snoop["payload"] = request.body